requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-apache-spark==4.11.3